hf_space/requirements.txt
# Hugging Face Space — models load from Hub, ZeroGPU for free GPU
spaces
gradio>=4.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
PyYAML>=6.0
numpy
safetensors
einops
xformers
